+++ Magritte/visitors/MAReaderWriterWalkerVisitors.py
